requirements.txt
django<2
httplib2
django-registration